Finished initial data generator
Okay this is pretty quick and dirty, but she works.

Executing this bad boy inserts random entries into the table.
Thanks to the use of the 'words.txt' file, the names of the
games are pretty amazing. The 'score' and 'budget' parameters
are explicitly made to correlate for the purpose of testing
learning algorithms.

Eventually I'll generify it more - maybe use a conf file for
arbitrary tables / keyspaces, maybe take command line arguments.
For my next trick, I just need to get it to execute several
hundred times.