Began work on cassandra linear regression example
regression.py will be an example of linear regression on
the example cassandra data that the generator script creates.
I made a change in that script as well; I want the data to be
more biased in favor of high budgets. I'll probably make a
function or some more intelligent formula for adding bias down
the road.

The regression.py script so far just makes a scatter plot and
has a cost function. I need to convert my Octave gradient
descent formula into Python and then see if I can make it
plot a line.