Added functions for Postgres data generator in R
Not quite done but the structure is there. I am not sure about
the method for generating random character strings. Seemed kind
of slow in the REPL. Once I can get the function for determining
malignancy worked out, I'll give it a spin in a VM and see if
we don't need a different approach.