Added schema and stub for R script
We're going with the tumor example again. We'll start with
just one feature and we'll use Postgres instead of Cassandra.
The table description is committed here as well as a test
function for an R script to generate data. Basically we will
follow the same pattern as we did for Python and Cassandra