Finished basic data inserter
So this Just Works but it won't do for large scale inserts.
The call to 'randomStrings' is kind of slow. I notice that
it can generate a bunch of strings in a single call if need
be, so I'll probably just do that and use a for loop on the
insert query.