Changing course with Java experiment
I got rid of all of the school / student code. I stashed
it elsewhere and decided not to do another regression
example. I also want to get away from using simple data
generators. They are too artificial to be useful. Since
registering with Kaggle and completing the ML class, I
have gotten access to a number of better data sources.

So instead what I have decided to do with Java is to
create an example of Support Vector Machines. I will
make a class that will import the data from the text
files and insert it into MongoDB (working with mongo is
still a priority).

Once the data is inserted, we will create classes to read
and parse all of the JSON documents. From there another
class can implement the SVM algorithm.