Refactored profile.py and added functionality
We are going to continue outputting small csv's for each
table for the time being. This commit breaks up the code
into more functions and sets up some additional data
gathering. Specifically we have a method for getting the
table size on disk and a query for getting the fields
(columns).

Once we munge it all together, we'll have the schema, size,
and columns for every table in its own small csv. From there
we can export to a Google doc, do automated diffing, or
other fun stuff.