Fixes for initial visualization
Okay, after some hacking around I have a simple visualization
working for the Python credit default dataset.

It's just using a small number of rows to plot defaults
against borrower's age and education level. Not much
obvious correlation at first blush, though the correlations.py
script and the graph suggest that younger and less educated
people are more likely to default.

The purpose of this is just to get some code down on paper.
From here we'll try to implement a simple logistic regression
using these really bad features. Once we have working code
for that, we can start following the suggestions in the readme
to actually do something intelligent.